+++ src/core.mjs
@@ -1,18 +1,22 @@
+/* eslint-disable indent */
 import path from 'path'
 import { getConfig } from './config.mjs'
 import { applyColor } from './transform.mjs'
 import { TICK, CROSS, EXIT_CODES } from './constants.mjs'
 import { timeStamp, printExecutionTime } from './support.mjs'
+import * as assertions from './assertions.mjs'
+import { AssertionError } from './assertionError.mjs'
 
 const config = getConfig()
 
+let currentTest
 let successes = 0
 const failures = []
 let describeStack = []
 
 let hasBeforeAll = false
-let beforeAllStack = []
 let hasAfterAll = false
+let beforeAllStack = []
 let afterAllStack = []
 
 // Runner entry point
@@ -29,6 +33,18 @@ export const run = async () => {
   printExecutionTime(startTimeStamp, endTimeStamp)
   process.exit(failures.length > 0 ? EXIT_CODES.failures : EXIT_CODES.ok)
 }
+
+const createDescribe = (name) => ({
+  name,
+  beforeEach: [],
+  afterEach: [],
+})
+
+const createTest = (name) => ({
+  name,
+  errors: [],
+  describeStack,
+})
 
 const last = (arr) => arr[arr.length - 1]
 
@@ -88,29 +104,28 @@ const invokeAfterAll = () => {
   }
 }
 
-const makeDescribe = (name) => ({
-  name,
-  beforeEach: [],
-  afterEach: [],
-})
-
 export const describe = (name, body) => {
-  describeStack = [...describeStack, makeDescribe(name)]
+  describeStack = [...describeStack, createDescribe(name)]
   body()
   invokeAfterAll()
   describeStack = withoutLast(describeStack)
 }
 
 export const test = (name, body) => {
+  currentTest = createTest(name)
   try {
     invokeBeforeAll()
     invokeBeforeEach()
     body()
-    console.log(indent(applyColor(`  <green>${TICK}</green> ${name}`)))
-    successes++
   } catch (e) {
-    console.error(indent(applyColor(`  <red>${CROSS}</red> ${name}`)))
-    failures.push({ error: e, name, describeStack })
+    currentTest.errors.push(e)
+  }
+  if (currentTest.errors.length > 0) {
+    console.log(indent(applyColor(`  <red>${CROSS}</red> ${name}`)))
+    failures.push(currentTest)
+  } else {
+    successes++
+    console.log(indent(applyColor(`  <green>${TICK}</green> ${name}`)))
   }
   try {
     invokeAfterEach()
@@ -123,7 +138,9 @@ const indent = (message) => `${' '.repeat(describeStack.length * 2)}${message}`
 
 const printFailureMsg = (failure) => {
   console.error(applyColor(fullTestDescription(failure)))
-  console.error(failure.error)
+  failure.errors.forEach((error) => {
+    console.error(error)
+  })
   console.error('')
 }
 
@@ -149,3 +166,21 @@ const fullTestDescription = ({ name, describeStack }) =>
   [...describeStack, { name }]
     .map(({ name }) => `<bold>${name}</bold>`)
     .join(' → ')
+
+const assertionsHandler = (actual) => ({
+  get:
+    (_, name) =>
+    (...args) => {
+      try {
+        assertions[name](actual, ...args)
+      } catch (e) {
+        if (e instanceof AssertionError) {
+          currentTest.errors.push(e)
+        } else {
+          throw e
+        }
+      }
+    },
+})
+
+export const expect = (actual) => new Proxy({}, assertionsHandler(actual))