+++ src/runnerError.mjs
@@ -1,0 +1,6 @@
+export class RunnerError extends Error {
+  // eslint-disable-next-line space-before-function-paren, no-useless-constructor
+  constructor(message) {
+    super(message)
+  }
+}